Rename common components
