+++ client/src/components/_common/SubmitButton.css
@@ -1,8 +1,4 @@
 .submit-button {
-    margin-top: 1.5rem;
-}
-
-.submit-button__input {
     background-color: #f5a254;
     border: 0.05rem solid #696969;
     border-radius: 0.75rem;
@@ -12,13 +8,15 @@
     font-family: 'Libre Baskerville', sans-serif;
     font-size: 1rem;
     font-weight: bold;
+    height: 2.75rem;
+    margin-top: 1.5rem;
     letter-spacing: 0.1rem;
     padding: 0.75rem 3rem;
     transition: all 0.3s ease 0s;
     outline: none;
 }
 
-.submit-button__input:hover {
+.submit-button:hover {
     box-shadow: 0 0.75rem 1.5rem rgba(0, 0, 0, 0.3);
     transform: translateY(-0.25rem);
 }
