+++ client/src/components/_common/SubmitButton.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import './SubmitButton.css';
+
+const SubmitButton = ({ value }) => {
+    return (
+        <button className='submit-button' type='submit' value={value}>
+            {value}
+        </button>
+    );
+};
+
+export default SubmitButton;
