--- client/src/components/common/SubmitButton.js
@@ -1,17 +1,0 @@
-import React from 'react';
-
-import './SubmitButton.css';
-
-const SubmitButton = ({ value }) => {
-    return (
-        <div className='submit-button'>
-            <input
-                className='submit-button__input'
-                type='submit'
-                value={value}
-            />
-        </div>
-    );
-};
-
-export default SubmitButton;
